requirements.txt
streamlit
faiss-cpu
langchain
python-dotenv
google-generativeai
PyMuPDF
python-docx
tqdm
sentence-transformers